chore: remove duplicate monolithic test files after split (#201)
Delete original test files that were not removed when tests were split
into focused modules. Two tests missing from split files were migrated
to test_stage3_image_conversion.py before deletion.
